fix: ErrorStatistics 필드 타입 변경

+++ src/main/java/S13P31A306/loglens/domain/dashboard/dto/opensearch/ErrorStatistics.java
@@ -4,7 +4,7 @@ package S13P31A306.loglens.domain.dashboard.dto.opensearch;
  * 에러 통계 정보
  */
 public record ErrorStatistics(
-        Long totalErrors,
+        Integer totalErrors,
         Integer uniqueErrorTypes
 ) {
 }
